+++ src/Data/Poly/Internal/Dense/Field.hs
@@ -1,12 +1,13 @@
 -- |
--- Module:      Data.Poly.Internal.Dense.Fractional
+-- Module:      Data.Poly.Internal.Dense.Field
 -- Copyright:   (c) 2019 Andrew Lelechenko
 -- Licence:     BSD3
 -- Maintainer:  Andrew Lelechenko <andrew.lelechenko@gmail.com>
 --
--- GcdDomain for Fractional underlying
+-- GcdDomain for Field underlying
 --
 
+{-# LANGUAGE ConstraintKinds            #-}
 {-# LANGUAGE CPP                        #-}
 {-# LANGUAGE FlexibleInstances          #-}
 {-# LANGUAGE PatternSynonyms            #-}
@@ -17,37 +18,44 @@
 
 #if MIN_VERSION_semirings(0,4,2)
 
-module Data.Poly.Internal.Dense.Fractional
-  ( fractionalGcd
+module Data.Poly.Internal.Dense.Field
+  ( fieldGcd
   , gcdExt
   ) where
 
-import Prelude hiding (quot, rem, gcd)
+import Prelude hiding (quotRem, quot, rem, gcd)
 import Control.Exception
 import Control.Monad
 import Control.Monad.Primitive
 import Control.Monad.ST
 import Data.Euclidean
+#if !MIN_VERSION_semirings(0,5,0)
 import Data.Semiring (Ring)
+#endif
+import Data.Semiring (times, minus, zero, one)
 import qualified Data.Vector.Generic as G
 import qualified Data.Vector.Generic.Mutable as MG
 
 import Data.Poly.Internal.Dense
 import Data.Poly.Internal.Dense.GcdDomain ()
 
-instance (Eq a, Eq (v a), Ring a, GcdDomain a, Fractional a, G.Vector v a) => Euclidean (Poly v a) where
+#if !MIN_VERSION_semirings(0,5,0)
+type Field a = (Euclidean a, Ring a, Fractional a)
+#endif
+
+instance (Eq a, Eq (v a), Field a, G.Vector v a) => Euclidean (Poly v a) where
   degree (Poly xs) = fromIntegral (G.length xs)
 
-  quotRem (Poly xs) (Poly ys) = (toPoly qs, toPoly rs)
+  quotRem (Poly xs) (Poly ys) = (toPoly' qs, toPoly' rs)
     where
       (qs, rs) = quotientAndRemainder xs ys
   {-# INLINE quotRem #-}
 
-  rem (Poly xs) (Poly ys) = toPoly $ remainder xs ys
+  rem (Poly xs) (Poly ys) = toPoly' $ remainder xs ys
   {-# INLINE rem #-}
 
 quotientAndRemainder
-  :: (Fractional a, G.Vector v a)
+  :: (Field a, G.Vector v a)
   => v a
   -> v a
   -> (v a, v a)
@@ -63,16 +71,16 @@ quotientAndRemainder xs ys
     G.unsafeCopy rs xs
     forM_ [lenQs - 1, lenQs - 2 .. 0] $ \i -> do
       r <- MG.unsafeRead rs (lenYs - 1 + i)
-      let q = r / G.unsafeLast ys
+      let q = r `quot` G.unsafeLast ys
       MG.unsafeWrite qs i q
       forM_ [0 .. lenYs - 1] $ \k -> do
-        MG.unsafeModify rs (\c -> c - q * G.unsafeIndex ys k) (i + k)
+        MG.unsafeModify rs (\c -> c `minus` q `times` G.unsafeIndex ys k) (i + k)
     let rs' = MG.unsafeSlice 0 lenYs rs
     (,) <$> G.unsafeFreeze qs <*> G.unsafeFreeze rs'
 {-# INLINE quotientAndRemainder #-}
 
 remainder
-  :: (Fractional a, G.Vector v a)
+  :: (Field a, G.Vector v a)
   => v a
   -> v a
   -> v a
@@ -86,7 +94,7 @@ remainder xs ys
 {-# INLINE remainder #-}
 
 remainderM
-  :: (PrimMonad m, Fractional a, G.Vector v a)
+  :: (PrimMonad m, Field a, G.Vector v a)
   => G.Mutable v (PrimState m) a
   -> G.Mutable v (PrimState m) a
   -> m ()
@@ -104,58 +112,58 @@ remainderM xs ys
         y <- MG.unsafeRead ys k
         -- do not move r / yLast outside the loop,
         -- because of numerical instability
-        MG.unsafeModify xs (\c -> c - r * y / yLast) (i + k)
+        MG.unsafeModify xs (\c -> c `minus` r `times` y `quot` yLast) (i + k)
 {-# INLINE remainderM #-}
 
-fractionalGcd
-  :: (Eq a, Fractional a, G.Vector v a)
+fieldGcd
+  :: (Eq a, Field a, G.Vector v a)
   => Poly v a
   -> Poly v a
   -> Poly v a
-fractionalGcd (Poly xs) (Poly ys) = toPoly $ runST $ do
+fieldGcd (Poly xs) (Poly ys) = toPoly' $ runST $ do
   xs' <- G.thaw xs
   ys' <- G.thaw ys
   gcdM xs' ys'
-{-# INLINE fractionalGcd #-}
+{-# INLINE fieldGcd #-}
 
 gcdM
-  :: (PrimMonad m, Eq a, Fractional a, G.Vector v a)
+  :: (PrimMonad m, Eq a, Field a, G.Vector v a)
   => G.Mutable v (PrimState m) a
   -> G.Mutable v (PrimState m) a
   -> m (v a)
 gcdM xs ys = do
-  ys' <- dropWhileEndM (== 0) ys
+  ys' <- dropWhileEndM (== zero) ys
   if MG.null ys' then G.unsafeFreeze xs else do
     remainderM xs ys'
     gcdM ys' xs
 {-# INLINE gcdM #-}
 
 -- | Execute the extended Euclidean algorithm.
--- For polynomials 'a' and 'b', compute their unique greatest common divisor 'g'
--- and the unique coefficient polynomial 's' satisfying 'a''s' + 'b''t' = 'g',
--- such that either 'g' is monic, or 'g = 0' and 's' is monic, or 'g = s = 0'.
+-- For polynomials @a@ and @b@, compute their unique greatest common divisor @g@
+-- and the unique coefficient polynomial @s@ satisfying @as + bt = g@,
+-- such that either @g@ is monic, or @g = 0@ and @s@ is monic, or @g = s = 0@.
 --
 -- >>> gcdExt (X^2 + 1 :: UPoly Double) (X^3 + 3 * X :: UPoly Double)
 -- (1.0, 0.5 * X^2 + (-0.0) * X + 1.0)
 -- >>> gcdExt (X^3 + 3 * X :: UPoly Double) (3 * X^4 + 3 * X^2 :: UPoly Double)
 -- (1.0 * X + 0.0,(-0.16666666666666666) * X^2 + (-0.0) * X + 0.3333333333333333)
 gcdExt
-  :: (Eq a, Fractional a, GcdDomain a, Ring a, G.Vector v a, Eq (v a))
+  :: (Eq a, Field a, G.Vector v a, Eq (v a))
   => Poly v a
   -> Poly v a
   -> (Poly v a, Poly v a)
 gcdExt xs ys = case scaleMonic gs of
-  Just (c', gs') -> (gs', scale 0 c' ss)
+  Just (c', gs') -> (gs', scale' zero c' ss)
   Nothing -> case scaleMonic ss of
-    Just (_, ss') -> (0, ss')
-    Nothing -> (0, 0)
+    Just (_, ss') -> (zero, ss')
+    Nothing -> (zero, zero)
   where
-    (gs, ss) = go ys xs 0 1
+    (gs, ss) = go ys xs zero one
       where
-        go r r' s s'
-          | r == 0 = (r', s')
-          | otherwise = case r' `quot` r of
-            q -> go (r' - q * r) r (s' - q * s) s
+        go r' r s' s
+          | r' == zero = (r, s)
+          | otherwise  = case r `quotRem` r' of
+            (q, r'') -> go r'' r' (s `minus` q `times` s') s'
 {-# INLINE gcdExt #-}
 
 -- | Scale a non-zero polynomial such that its leading coefficient is one,
@@ -166,16 +174,16 @@ gcdExt xs ys = case scaleMonic gs of
 -- >>> scaleMonic (3 * X^4 + 3 * X^2 :: UPoly Double)
 -- Just (0.3333333333333333, 1.0 * X^4 + 0.0 * X^3 + 1.0 * X^2 + 0.0 * X + 0.0)
 scaleMonic
-  :: (Eq a, Fractional a, GcdDomain a, Ring a, G.Vector v a, Eq (v a))
+  :: (Eq a, Field a, G.Vector v a, Eq (v a))
   => Poly v a
   -> Maybe (a, Poly v a)
 scaleMonic xs = case leading xs of
   Nothing -> Nothing
-  Just (_, c) -> let c' = recip c in Just (c', scale 0 c' xs)
+  Just (_, c) -> let c' = one `quot` c in Just (c', scale' zero c' xs)
 {-# INLINE scaleMonic #-}
 
 #else
 
-module Data.Poly.Internal.Dense.Fractional () where
+module Data.Poly.Internal.Dense.Field () where
 
 #endif