+++ src/Data/Poly/Semiring.hs
@@ -23,10 +23,16 @@ module Data.Poly.Semiring
   , pattern X
   , eval
   , deriv
+#if MIN_VERSION_semirings(0,5,0)
+  , integral
+#endif
 #if MIN_VERSION_semirings(0,4,2)
-  -- * Fractional coefficients
-  , PolyOverFractional(..)
+  -- * Polynomials over 'Field'
+  , PolyOverField(..)
   , gcdExt
+  , PolyOverFractional
+  , pattern PolyOverFractional
+  , unPolyOverFractional
 #endif
   ) where
 
@@ -36,9 +42,12 @@ import qualified Data.Vector.Generic as G
 import Data.Poly.Internal.Dense (Poly(..), VPoly, UPoly, leading)
 import qualified Data.Poly.Internal.Dense as Dense
 #if MIN_VERSION_semirings(0,4,2)
-import Data.Poly.Internal.Dense.Fractional (gcdExt)
+import Data.Poly.Internal.Dense.Field (gcdExt)
 import Data.Poly.Internal.Dense.GcdDomain ()
-import Data.Poly.Internal.PolyOverFractional
+import Data.Poly.Internal.PolyOverField
+#endif
+#if MIN_VERSION_semirings(0,5,0)
+import Data.Euclidean (Field)
 #endif
 
 -- | Make 'Poly' from a vector of coefficients
@@ -82,3 +91,13 @@ eval = Dense.eval'
 -- 3 * X^2 + 0 * X + 3
 deriv :: (Eq a, Semiring a, G.Vector v a) => Poly v a -> Poly v a
 deriv = Dense.deriv'
+
+#if MIN_VERSION_semirings(0,5,0)
+-- | Compute an indefinite integral of a polynomial,
+-- setting constant term to zero.
+--
+-- >>> integral (3 * X^2 + 3) :: UPoly Double
+-- 1.0 * X^3 + 0.0 * X^2 + 3.0 * X + 0.0
+integral :: (Eq a, Field a, G.Vector v a) => Poly v a -> Poly v a
+integral = Dense.integral'
+#endif