+++ src/Data/Poly/Sparse/Semiring.hs
@@ -7,8 +7,9 @@
 -- Sparse polynomials with 'Semiring' instance.
 --
 
-{-# LANGUAGE FlexibleContexts    #-}
-{-# LANGUAGE PatternSynonyms     #-}
+{-# LANGUAGE CPP              #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE PatternSynonyms  #-}
 
 module Data.Poly.Sparse.Semiring
   ( Poly
@@ -23,6 +24,13 @@ module Data.Poly.Sparse.Semiring
   , pattern X
   , eval
   , deriv
+#if MIN_VERSION_semirings(0,5,0)
+  , integral
+#endif
+#if MIN_VERSION_semirings(0,4,2)
+  -- * Polynomials over 'Field'
+  , gcdExt
+#endif
   ) where
 
 import Data.Semiring (Semiring)
@@ -30,8 +38,13 @@ import qualified Data.Vector.Generic as G
 
 import Data.Poly.Internal.Sparse (Poly(..), VPoly, UPoly, leading)
 import qualified Data.Poly.Internal.Sparse as Sparse
-import Data.Poly.Internal.Sparse.Fractional ()
+#if MIN_VERSION_semirings(0,4,2)
+import Data.Poly.Internal.Sparse.Field (gcdExt)
 import Data.Poly.Internal.Sparse.GcdDomain ()
+#endif
+#if MIN_VERSION_semirings(0,5,0)
+import Data.Euclidean (Field)
+#endif
 
 -- | Make 'Poly' from a list of (power, coefficient) pairs.
 -- (first element corresponds to a constant term).
@@ -74,3 +87,13 @@ eval = Sparse.eval'
 -- 3 * X^2 + 3
 deriv :: (Eq a, Semiring a, G.Vector v (Word, a)) => Poly v a -> Poly v a
 deriv = Sparse.deriv'
+
+#if MIN_VERSION_semirings(0,5,0)
+-- | Compute an indefinite integral of a polynomial,
+-- setting constant term to zero.
+--
+-- >>> integral (3 * X^2 + 3) :: UPoly Double
+-- 1.0 * X^3 + 3.0 * X
+integral :: (Eq a, Field a, G.Vector v (Word, a)) => Poly v a -> Poly v a
+integral = Sparse.integral'
+#endif