+++ test/Sparse.hs
@@ -11,13 +11,14 @@ module Sparse
   ( testSuite
   ) where
 
-import Prelude hiding (quotRem)
+import Prelude hiding (gcd, quotRem, rem)
 #if MIN_VERSION_semirings(0,4,2)
 import Data.Euclidean
 #endif
 import Data.Function
 import Data.Int
 import Data.List
+import Data.Maybe
 import Data.Poly.Sparse
 import qualified Data.Poly.Sparse.Semiring as S
 import Data.Proxy
@@ -57,6 +58,9 @@ testSuite = testGroup "Sparse"
     , lawsTests
     , evalTests
     , derivTests
+#if MIN_VERSION_semirings(0,4,2)
+    , gcdExtTests
+#endif
     ]
 
 lawsTests :: TestTree
@@ -248,6 +252,8 @@ derivTests :: TestTree
 derivTests = testGroup "deriv"
   [ testProperty "deriv = S.deriv" $
     \(p :: Poly V.Vector Integer) -> deriv p === S.deriv p
+  , testProperty "integral = S.integral" $
+    \(p :: Poly V.Vector Rational) -> integral p === S.integral p
   , testProperty "deriv . integral = id" $
     \(p :: Poly V.Vector Rational) -> deriv (integral p) === p
   , testProperty "deriv c = 0" $
@@ -264,3 +270,22 @@ derivTests = testGroup "deriv"
   --     deriv (eval (toPoly $ fmap (fmap $ monomial 0) $ unPoly p) q) ===
   --       deriv q * eval (toPoly $ fmap (fmap $ monomial 0) $ unPoly $ deriv p) q
   ]
+
+#if MIN_VERSION_semirings(0,4,2)
+gcdExtTests :: TestTree
+gcdExtTests = localOption (QuickCheckMaxSize 12) $ testGroup "gcdExt"
+  [ testProperty "gcdExt == S.gcdExt" $
+    \(a :: Poly V.Vector Rational) b ->
+      gcdExt a b === S.gcdExt a b
+  , testProperty "g == as (mod b) for gcdExt" $
+    \(a :: Poly V.Vector Rational) b -> b /= 0 ==>
+      uncurry ((. flip rem b) . (===) . flip rem b) ((* a) <$> gcdExt a b)
+  , testProperty "fst . gcdExt == gcd (mod units)" $
+    \(a :: Poly V.Vector Rational) b ->
+      fst (gcdExt a b) `sameUpToUnits` gcd a b
+  ]
+
+sameUpToUnits :: (Eq a, GcdDomain a) => a -> a -> Bool
+sameUpToUnits x y = x == y ||
+  isJust (x `divide` y) && isJust (y `divide` x)
+#endif