+++ changelog.md
@@ -1,3 +1,10 @@
+# 0.3.2.0
+
+* Add `NFData` instance.
+* Implement extended GCD.
+* Rename `PolyOverFractional` to `PolyOverField`.
+* Add `integral` with `Semiring`-based interface.
+
 # 0.3.1.0
 
 * Implement Karatsuba multiplication.